pyproject.toml
[tool.poetry]
name = "simple-webfinger"
version = "0.1.0"
description = "A simple, Flask-based webfinger handler."
authors = ["Andrew Williams <andy@tensixtyone.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
PyYAML = "^6.0.2"
uwsgi = "^2.0.26"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.6.0"
pytest-mock = "^3.12.0"
pytest-flask = "^1.3.0"

[tool.poetry.group.github]
optional = true

[tool.poetry.group.github.dependencies]
pytest-github-actions-annotate-failures = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
